+++ CHANGELOG.md
@@ -10,5 +10,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release of shiny-treeview
 - Select from hierarchical data in your [Shiny for Python](https://shiny.posit.co/py/) apps via `input_treeview()`
-- Create hierarchical data with nested `TreeItem` objects
+- Create hierarchical data with nested `TreeItem` objects.
+    - Parameters: `id`, `label`, `caption`, `children`, `disabled`.
 - Convert flat to hierarchical data with helper functions: `stratify_by_parent()`